requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
Django==4.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.14.0
idna==3.6
inflection==0.5.1
packaging==24.1
phonenumberslite==8.13.34
Pillow==10.4.0
psycopg2==2.9.10  # Updated to latest source version
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
whitenoise==6.6.0
wheel==0.43.0
dj-database-url==2.1.0
setuptools==70.0.0requests
